requirements.txt
accelerate==1.7.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
airportsdata==20250523
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astor==0.8.1
asttokens==3.0.1
async-lru==2.2.0
attrs==25.3.0
audioread==3.0.1
autoawq==0.2.9
babel==2.18.0
beautifulsoup4==4.13.4
blake3==1.0.5
bleach==6.3.0
cachetools==6.0.0
cbor==1.0.0
certifi==2025.4.26
cffi==2.0.0
charset-normalizer==3.4.2
click==8.2.1
cloudpickle==3.1.1
coloredlogs==15.0.1
comm==0.2.3
compressed-tensors==0.9.4
contourpy==1.3.2
cryptography==46.0.5
cupy-cuda12x==13.6.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==3.6.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
einops==0.8.1
email_validator==2.2.0
executing==2.2.1
faiss-cpu==1.11.0
fastapi==0.115.12
fastapi-cli==0.0.7
fastjsonschema==2.21.2
fastrlock==0.8.3
ffmpy==0.5.0
filelock==3.18.0
fire==0.7.0
FlagEmbedding==1.3.4
flatbuffers==25.2.10
fonttools==4.58.1
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.0
gekko==1.3.0
gguf==0.17.0
googleapis-common-protos==1.70.0
gradio==5.31.0
gradio_client==1.10.1
greenlet==3.2.2
groovy==0.1.2
grpcio==1.67.1
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.32.3
humanfriendly==10.0
idna==3.10
ijson==3.4.0
importlib_metadata==8.6.1
inscriptis==2.6.0
interegular==0.3.3
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
ir_datasets==0.5.10
isoduration==20.11.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.0
json5==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter_client==8.8.0
jupyter-console==6.6.3
jupyter_core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.8
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.63
langchain-openai==0.3.18
langchain-text-splitters==0.3.8
langsmith==0.3.42
lark==1.2.2
lazy_loader==0.4
llguidance==0.7.26
llvmlite==0.44.0
lm-format-enforcer==0.10.11
loguru==0.7.3
lxml==5.4.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.3
matplotlib-inline==0.2.1
mdurl==0.1.2
milvus-lite==2.4.12
mistral_common==1.5.6
mistune==3.2.0
modelscope==1.26.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.4.4
multiprocess==0.70.16
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
notebook==7.5.5
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.82.0
opencv-python-headless==4.11.0.86
opentelemetry-api==1.33.1
opentelemetry-exporter-otlp==1.33.1
opentelemetry-exporter-otlp-proto-common==1.33.1
opentelemetry-exporter-otlp-proto-grpc==1.33.1
opentelemetry-exporter-otlp-proto-http==1.33.1
opentelemetry-proto==1.33.1
opentelemetry-sdk==1.33.1
opentelemetry-semantic-conventions==0.54b1
opentelemetry-semantic-conventions-ai==0.4.9
optimum==1.25.3
orjson==3.10.18
outlines==0.1.11
outlines_core==0.1.26
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.6
partial-json-parser==0.2.1.1.post5
pdfminer.six==20251230
pdfplumber==0.11.9
peft==0.15.2
pexpect==4.9.0
pillow==11.2.1
pip==25.2
platformdirs==4.3.8
pooch==1.8.2
prometheus_client==0.22.0
prometheus-fastapi-instrumentator==7.1.0
prompt_toolkit==3.0.52
propcache==0.3.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.9.1
pydub==0.25.1
Pygments==2.19.1
pymilvus==2.5.9
pymilvus.model==0.3.2
pymongo==4.13.0
PyMuPDF==1.25.5
pyparsing==3.2.3
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
ragas==0.2.15
rank-bm25==0.2.2
ray==2.46.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-toolkit==0.14.7
rouge==1.0.1
rouge-chinese==1.0.3
rpds-py==0.25.1
ruff==0.11.12
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
semantic-version==2.10.0
Send2Trash==2.1.0
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==78.1.1
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
sse-starlette==2.3.5
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
termcolor==3.1.0
terminado==0.18.1
text2vec==1.3.5
threadpoolctl==3.6.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.1
tomlkit==0.13.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.5.5
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.1
trec-car-tools==2.6
triton==3.3.0
trl==0.9.6
typer==0.16.0
typing_extensions==4.13.2
typing-inspect==0.9.0
typing-inspection==0.4.0
tyro==0.8.14
tzdata==2025.2
ujson==5.10.0
unlzw3==0.2.3
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
vllm==0.9.0.1
warc3-wet==0.2.5
warc3-wet-clueweb09==0.2.5
watchfiles==1.0.5
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
wheel==0.45.1
widgetsnbextension==4.0.15
wrapt==1.17.2
xformers==0.0.30
xgrammar==0.1.19
xxhash==3.5.0
yarl==1.20.0
zipp==3.22.0
zlib-state==0.1.9
zstandard==0.23.0
